fix: tests for https://github.com/mongodb/specifications/pull/1377
